Moved files to have a correct test-folder structure. Added and renamed constants for consistency.
