+++ src/test/java/no/ntnu/mycbr/rest/common/CommonConstants.java
@@ -1,0 +1,17 @@
+package no.ntnu.mycbr.rest.common;
+
+public class CommonConstants {
+
+    // Path variable for concept
+    public static final String PATH_CONCEPT = "/concepts/";
+
+    // Path variable for casebase
+    public static final String PATH_CASEBASE = "/casebases/";
+
+    // Path variable for instances
+    public static final String PATH_INSTANCES = "/instances/";
+
+    // Root and two steps into the JSON file.
+    public static final String JSON_PATH = "$..";
+
+}
