+++ src/test/java/no/ntnu/mycbr/rest/controller/InstanceControllerIntegrationTest.java
@@ -1,4 +1,4 @@
-package no.ntnu.mycbr.rest.test;
+package no.ntnu.mycbr.rest.controller;
 
 import no.ntnu.mycbr.rest.App;
 import no.ntnu.mycbr.rest.service.CaseBaseService;
@@ -7,7 +7,6 @@ import no.ntnu.mycbr.rest.service.InstanceService;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
-import org.hamcrest.Matchers;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -26,10 +25,9 @@ import java.util.*;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
-import no.ntnu.mycbr.rest.test.utils.TestSetup;
-import static no.ntnu.mycbr.rest.test.utils.TestSetup.*;
-import no.ntnu.mycbr.rest.test.utils.CommonConstants;
-import static no.ntnu.mycbr.rest.test.utils.CommonConstants.*;
+import no.ntnu.mycbr.rest.controller.utils.TestSetup;
+import static no.ntnu.mycbr.rest.controller.utils.TestSetup.*;
+import static no.ntnu.mycbr.rest.common.CommonConstants.*;
 
 @RunWith(SpringRunner.class)
 //@WebMvcTest(ConceptController.class)
@@ -62,22 +60,22 @@ public class InstanceControllerIntegrationTest {
     @Test
     public void addInstancesJSON() throws Exception {
         // add an instance
-        mockMvc.perform(post(path_concept + CONCEPT_NAME + path_casebase + CASE_BASE_NAME + path_instances)
+        mockMvc.perform(post(PATH_CONCEPT + CONCEPT_NAME + PATH_CASEBASE + CASE_BASE_NAME + PATH_INSTANCES)
                 .param("cases", INSTANCE)
                 .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk());
         // check if it exists(last ".andExpect" in the code under might be wrong, but method itself does not work)
-        mockMvc.perform(get(path_concept+CONCEPT_NAME+path_instances)
+        mockMvc.perform(get(PATH_CONCEPT +CONCEPT_NAME+ PATH_INSTANCES)
                 .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
-                .andExpect((jsonPath("$.."+ATT_DOUBLE_1).value("0.3")));
+                .andExpect((jsonPath(JSON_PATH+ATT_DOUBLE_1).value("0.3")));
         logger.debug("results..");
 
     }
     @Test
     public void addInstanceJSONTest() throws Exception {
         // add an instance
-        mockMvc.perform(put(path_concept + CONCEPT_NAME + path_casebase + CASE_BASE_NAME + path_instances + TEST_CASE_ID)
+        mockMvc.perform(put(PATH_CONCEPT + CONCEPT_NAME + PATH_CASEBASE + CASE_BASE_NAME + PATH_INSTANCES + TEST_CASE_ID)
                 .param("casedata", "{" +
                        "\"" + ATT_DOUBLE_1 +"\""+ ":0.7,"+
                        "\"" + ATT_DOUBLE_2 +"\""+ ":0.4"+
@@ -86,18 +84,18 @@ public class InstanceControllerIntegrationTest {
         .andExpect(status().isOk());
         // check if it exists
         // todo: use getInstance instead of getAllInstancesInCaseBase
-        mockMvc.perform(get(path_concept + CONCEPT_NAME + path_casebase + CASE_BASE_NAME + path_instances)
+        mockMvc.perform(get(PATH_CONCEPT + CONCEPT_NAME + PATH_CASEBASE + CASE_BASE_NAME + PATH_INSTANCES)
                 .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
-                .andExpect((jsonPath("$.."+ATT_DOUBLE_1).value("0.7")))
-                .andExpect((jsonPath("$.."+ATT_DOUBLE_2).value("0.4")))
-                .andExpect((jsonPath("$..caseID").value("car1"))
+                .andExpect((jsonPath(JSON_PATH+ATT_DOUBLE_1).value("0.7")))
+                .andExpect((jsonPath(JSON_PATH+ATT_DOUBLE_2).value("0.4")))
+                .andExpect((jsonPath(JSON_PATH+"caseID").value("car1"))
         );
     }
     @Test
     public void deleteInstanceTest() throws Exception {
         // add an instance: "car1"
-        mockMvc.perform(put(path_concept + CONCEPT_NAME + path_casebase + CASE_BASE_NAME + path_instances + InstanceControllerIntegrationTest.TEST_CASE_ID)
+        mockMvc.perform(put(PATH_CONCEPT + CONCEPT_NAME + PATH_CASEBASE + CASE_BASE_NAME + PATH_INSTANCES + TEST_CASE_ID)
                 .param("casedata", "{" +
                         "\"" + ATT_DOUBLE_1 +"\""+ ":0.7,"+
                         "\"" + ATT_DOUBLE_2 +"\""+ ":0.4"+
@@ -105,7 +103,7 @@ public class InstanceControllerIntegrationTest {
                 .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk());
         // delete the instance: "car1"
-        mockMvc.perform(delete(path_concept + CONCEPT_NAME + path_casebase + CASE_BASE_NAME+ path_instances + InstanceControllerIntegrationTest.TEST_CASE_ID)
+        mockMvc.perform(delete(PATH_CONCEPT + CONCEPT_NAME + PATH_CASEBASE + CASE_BASE_NAME+ PATH_INSTANCES + TEST_CASE_ID)
                 .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
                 .andExpect(content().string("true"));
