--- src/test/java/no/ntnu/mycbr/rest/test/utils/CommonConstants.java
@@ -1,8 +1,0 @@
-package no.ntnu.mycbr.rest.test.utils;
-
-public class CommonConstants {
-
-    public static final String path_concept = "/concepts/";
-    public static final String path_casebase = "/casebases/";
-    public static final String path_instances = "/instances/";
-}
